# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Supercheck API",
    "version": "1.0.0",
    "description": "API for test automation, monitoring, and CI/CD integration. Supercheck provides Playwright browser testing, k6 performance testing, uptime monitoring, and alerting — all managed via API or CLI.",
    "contact": {
      "name": "Supercheck",
      "url": "https://supercheck.io"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://github.com/supercheck-io/supercheck/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://app.supercheck.io",
      "description": "Production (Cloud)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/jobs/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getJob",
        "summary": "Get job details",
        "tags": [
          "Jobs"
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateJob",
        "summary": "Update a job",
        "description": "Update job configuration. Only provided fields are changed.",
        "tags": [
          "Jobs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "cronSchedule": {
                    "type": "string"
                  },
                  "timeoutSeconds": {
                    "type": "integer"
                  },
                  "retryCount": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "paused"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteJob",
        "summary": "Delete a job",
        "tags": [
          "Jobs"
        ],
        "responses": {
          "200": {
            "description": "Job deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/runs": {
      "get": {
        "operationId": "listRuns",
        "summary": "List runs",
        "description": "List job runs with optional filters for job ID and status.",
        "tags": [
          "Runs"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "name": "jobId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by job ID"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "running",
                "completed",
                "failed",
                "cancelled"
              ]
            },
            "description": "Filter by status"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of runs",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Run"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/runs/{runId}": {
      "parameters": [
        {
          "name": "runId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getRun",
        "summary": "Get run details",
        "tags": [
          "Runs"
        ],
        "responses": {
          "200": {
            "description": "Run details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteRun",
        "summary": "Delete a run",
        "tags": [
          "Runs"
        ],
        "responses": {
          "200": {
            "description": "Run deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CLI token (`sck_live_*`) for full API access, or trigger key (`sck_trigger_*`) for job trigger endpoint only."
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        },
        "description": "Page number for pagination"
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 50,
          "minimum": 1,
          "maximum": 100
        },
        "description": "Items per page"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Not found"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPrevPage": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "running"
            ]
          },
          "cronSchedule": {
            "type": "string",
            "nullable": true
          },
          "timeoutSeconds": {
            "type": "integer"
          },
          "retryCount": {
            "type": "integer"
          },
          "lastRunAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "name": "Nightly E2E Suite",
          "description": "Run all browser tests nightly",
          "status": "active",
          "cronSchedule": "0 2 * * *",
          "timeoutSeconds": 300,
          "retryCount": 1,
          "lastRunAt": "2025-01-15T02:00:00Z",
          "createdAt": "2025-01-01T10:00:00Z",
          "updatedAt": "2025-01-15T02:00:00Z"
        }
      },
      "Run": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "jobName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "failed",
              "cancelled"
            ]
          },
          "trigger": {
            "type": "string",
            "enum": [
              "manual",
              "scheduled",
              "api",
              "ci"
            ]
          },
          "duration": {
            "type": "integer",
            "description": "Duration in milliseconds",
            "nullable": true
          },
          "result": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "example": {
          "id": "f1e2d3c4-b5a6-7890-abcd-ef1234567890",
          "jobId": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "jobName": "Nightly E2E Suite",
          "status": "completed",
          "trigger": "scheduled",
          "duration": 45200,
          "result": {
            "passed": 12,
            "failed": 0,
            "skipped": 1
          },
          "createdAt": "2025-01-15T02:00:00Z",
          "completedAt": "2025-01-15T02:00:45Z"
        }
      },
...
    }
  }
}